requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
Flask-JWT-Extended==4.4.0
Flask-MySQLdb==0.2.0
Werkzeug==2.2.3
